Move `Signal` operators to files
