+++ Source/Reactive/Signal Operators/AsVoid.swift
@@ -1,0 +1,3 @@
+public extension SignalType {
+	var asVoid: Signal<Void> { map { _ in () } }
+}
