+++ Source/Reactive/Signal Operators/CompactMap.swift
@@ -1,0 +1,8 @@
+public extension SignalType {
+	func compactMap<B>(_ f: @escaping (A) -> B?) -> Signal<B> {
+		self
+			.map(f)
+			.filter { $0 != nil }
+			.map { $0! }
+	}
+}
