+++ Source/Reactive/Signal Operators/FlatMap.swift
@@ -1,0 +1,5 @@
+public extension SignalType {
+	func flatMap<B>(_ f: @escaping (A) -> Signal<B>) -> Signal<B> {
+		map(f).flatten()
+	}
+}
