+++ Source/Reactive/Signal Operators/Map.swift
@@ -1,0 +1,7 @@
+public extension SignalType {
+	func map<B>(_ f: @escaping (A) -> B) -> Signal<B> {
+		Signal<B> {
+			observe($0 • f)
+		}
+	}
+}
