+++ Source/Reactive/Signal Operators/Merge.swift
@@ -1,0 +1,10 @@
+public extension SignalType {
+	func merge(_ signal: Signal<A>) -> Signal<A> {
+		.init {
+			CompositeDisposable([
+				observe($0),
+				signal.observe($0)
+			])
+		}
+	}
+}
