+++ Source/Reactive/Signal Operators/Observe+ExecutionContext.swift
@@ -1,0 +1,5 @@
+public extension SignalType {
+	func observe(_ ctx: ExecutionContext, _ f: @escaping (A) -> Void) -> Disposable {
+		observe { x in ctx.run { f(x) } }
+	}
+}
